Initial work to re-write linter
