+++ extension/server/src/providers/linter/index.ts
@@ -3,7 +3,7 @@ import { CodeAction, CodeActionKind, Diagnostic, DiagnosticSeverity, DidChangeWa
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import { URI } from 'vscode-uri';
 import { documents, parser } from '..';
-import { IssueRange, Rules } from '../../../../../language';
+import { IssueRange, Rules } from '../../../../../language/parserTypes';
 import Linter from '../../../../../language/linter';
 import Cache from '../../../../../language/models/cache';
 import codeActionsProvider from './codeActions';
