+++ language/parser.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-case-declarations */
 
-import { createBlocks, parseStatement } from "./statement";
+import { createBlocks, tokenise } from "./tokens";
 
 import Cache from "./models/cache";
 import Declaration from "./models/declaration";
@@ -1280,11 +1280,11 @@ export default class Parser {
    * @param {string[]} parts 
    */
   static expandKeywords(parts) {
-    /** @type {import(".").Keywords} */
+    /** @type {import("./parserTypes").Keywords} */
     const keyvalues = {};
 
     if (parts.length > 0) {
-      const keywordParts = createBlocks(parseStatement(parts.join(` `)));
+      const keywordParts = createBlocks(tokenise(parts.join(` `)));
 
       for (let i = 0; i < keywordParts.length; i++) {
         if (keywordParts[i].value) {
