+++ language/statement.ts
@@ -1,0 +1,40 @@
+import { IRange, Token } from "./types";
+
+export default class Statement {
+	private beginBlock = false;
+
+  constructor(public tokens: Token[], public range: IRange, public indent: number = 0) {}
+
+	getTokenByOffset(offset: number) {
+		const blockSearch = (tokens: Token[]): Token|undefined => {
+			const token = tokens.find(token => offset >= token.range.start && offset <= token.range.end);
+			
+			if (token?.type === `block` && token.block) {
+				return blockSearch(token.block);
+			}
+
+			return token;
+		}
+
+		return blockSearch(this.tokens);
+	}
+
+	static trimTokens(tokens: Token[]) {
+    if (tokens.length > 0) {
+      let realFirstToken = tokens.findIndex(t => t.type !== `newline`);
+      if (realFirstToken < 0) realFirstToken = 0;
+
+      let realLastToken = 0;
+
+      for (let i = tokens.length - 1; i >= 0; i--) {
+        if (tokens[i].type !== `newline`) {
+          realLastToken = i + 1;
+          break;
+        }
+      }
+
+      tokens = tokens.slice(realFirstToken, realLastToken);
+    }
+    return tokens;
+  }
+}
