+++ language/types.ts
@@ -1,0 +1,14 @@
+
+
+export interface IRange {
+  line: number;
+  start: number;
+  end: number;
+}
+
+export interface Token {
+  value?: string;
+  block?: Token[];
+  type: string;
+  range: IRange;
+}
